last version

+++ src/main/java/com/example/gestionbacaire/Dao/AgenceDao.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.gestionbacaire.bean.Agence;
 
 public interface AgenceDao extends JpaRepository<Agence, Long> {
-  public Agence findByNumero(int numero);
+  public Agence findByNumero(Long numero);
 }
+
